Split the fs.py files in more focused files.
