+++ tests/test_dev_tools/components/test_save_new_version.py
@@ -1,7 +1,7 @@
 from tests.support.dirs.my_path import MyPath
 from tests.support.files import make_file
 from tests.support.tools.version_saver import VersionSaver
-from trashcli.fs import read_file
+from trashcli.fslib.fs_func import read_file
 from trashcli.put.fs.real_fs import RealFs
