+++ tests/test_put/components/test_persist_trashinfo.py
@@ -5,7 +5,7 @@ import unittest
 import pytest
 from six import StringIO
 
-from trashcli.fs import read_file
+from trashcli.fslib.fs_func import read_file
 from trashcli.put.fs.real_fs import RealFs
 from trashcli.put.janitor_tools.info_file_persister import InfoFilePersister
 from trashcli.put.janitor_tools.info_file_persister import TrashinfoData
