+++ tests/test_restore/cmd/test_end_to_end_restore.py
@@ -9,7 +9,7 @@ import pytest
 from tests.support.fakes.fake_trash_dir import FakeTrashDir
 from tests.support.dirs.my_path import MyPath
 from tests.support.run.run_command import run_command
-from trashcli.fs import read_file
+from trashcli.fslib.fs_func import read_file
 
 
 @pytest.mark.slow
