+++ tests/test_restore/components/trashed_files/test_trashed_files_integration.py
@@ -5,7 +5,7 @@ from tests.support.py2mock import Mock
 
 from tests.support.files import make_file, require_empty_dir
 from tests.support.dirs.remove_dir_if_exists import remove_dir_if_exists
-from trashcli.fs import remove_file
+from trashcli.fslib.fs_func import remove_file
 from trashcli.restore.file_system import RealFileReader
 from trashcli.restore.info_dir_searcher import InfoDirSearcher, FileFound
 from trashcli.restore.trashed_files import TrashedFiles
