+++ tests/test_rm/components/test_file_remover.py
@@ -1,6 +1,6 @@
 import unittest
 
-from trashcli.rm.file_remover import FileRemover
+from trashcli.rm.real_remover_fs import RealRemoverFs
 
 
 try:
@@ -11,6 +11,6 @@ except NameError:
 
 class TestFileRemover(unittest.TestCase):
     def test_remove_file_fails_when_file_does_not_exists(self):
-        file_remover = FileRemover()
+        file_remover = RealRemoverFs()
         self.assertRaises(FileNotFoundError, file_remover.remove_file2,
                           '/non/existing/path')
