+++ tests/test_rm/components/test_list_trash_info.py
@@ -4,7 +4,7 @@ import pytest
 
 from tests.support.fakes.fake_trash_dir import FakeTrashDir
 from tests.support.dirs.my_path import MyPath
-from trashcli.file_system_reader import FileSystemReader
+from trashcli.fslib.file_system_reader import FileSystemReader
 from trashcli.rm.list_trashinfo import ListTrashinfos
