+++ tests/test_support/test_filesystem.py
@@ -5,7 +5,7 @@ import unittest
 
 import pytest
 
-from trashcli.fs import has_sticky_bit, mkdirs, is_sticky_dir
+from trashcli.fslib.fs_func import has_sticky_bit, mkdirs, is_sticky_dir
 from tests.support.files import make_empty_file, set_sticky_bit, unset_sticky_bit
 from tests.support.dirs.my_path import MyPath
