+++ trashcli/empty/existing_file_remover.py
@@ -1,4 +1,4 @@
-from trashcli.fs import RealRemoveFileIfExists, RealRemoveFile2
+from trashcli.fslib.real_fs_operations import RealRemoveFile2, RealRemoveFileIfExists
 
 
 class ExistingFileRemover(RealRemoveFileIfExists, RealRemoveFile2):
