+++ trashcli/empty/file_system_dir_reader.py
@@ -1,4 +1,4 @@
-from trashcli.fs import RealExists, RealEntriesIfDirExists
+from trashcli.fslib.real_fs_operations import RealEntriesIfDirExists, RealExists
 from trashcli.lib.dir_reader import DirReader
