+++ trashcli/empty/top_trash_dir_rules_file_system_reader.py
@@ -1,4 +1,4 @@
-from trashcli.fs import RealExists, RealIsStickyDir, RealIsSymLink, \
+from trashcli.fslib.real_fs_operations import RealExists, RealIsStickyDir, RealIsSymLink, \
     RealIsWorldWritable
 from trashcli.trash_dirs_scanner import TopTrashDirRules
