+++ trashcli/fslib/file_system_reader.py
@@ -1,5 +1,5 @@
-from trashcli.fs import RealIsStickyDir, RealHasStickyBit, \
-    RealIsSymLink, RealContentsOf, RealEntriesIfDirExists, RealExists
+from trashcli.fslib.real_fs_operations import RealEntriesIfDirExists, RealExists, \
+    RealHasStickyBit, RealIsStickyDir, RealIsSymLink, RealContentsOf
 from trashcli.list.fs import FileSystemReaderForListCmd
