+++ trashcli/lib/dir_reader.py
@@ -2,8 +2,8 @@ from __future__ import absolute_import
 
 from trashcli.compat import Protocol
 
-from trashcli.fs import EntriesIfDirExists, PathExists, RealEntriesIfDirExists, \
-    RealExists
+from trashcli.fslib.fs_operations import EntriesIfDirExists, PathExists
+from trashcli.fslib.real_fs_operations import RealEntriesIfDirExists, RealExists
 
 
 class DirReader(
