+++ trashcli/list/main.py
@@ -4,8 +4,8 @@ import sys
 
 import trashcli.trash
 from trashcli.empty.main import ContentReader
-from trashcli.file_system_reader import FileSystemReader
-from trashcli.fs import RealContentsOf
+from trashcli.fslib.file_system_reader import FileSystemReader
+from trashcli.fslib.real_fs_operations import RealContentsOf
 from trashcli.fstab.volume_listing import RealVolumesListing
 from trashcli.fstab.real_volume_of import RealVolumeOf
 from trashcli.fstab.volume_of import VolumeOf
