+++ trashcli/rm/real_remover_fs.py
@@ -1,0 +1,6 @@
+from trashcli.fslib.real_fs_operations import RealRemoveFile2, \
+    RealRemoveFileIfExists
+
+
+class RealRemoverFs(RealRemoveFile2, RealRemoveFileIfExists):
+    pass
