+++ trashcli/trash_dirs_scanner.py
@@ -2,7 +2,7 @@ import os
 from typing import Iterable
 from trashcli.compat import Protocol
 
-from trashcli.fs import PathExists, IsStickyDir, IsSymLink
+from trashcli.fslib.fs_operations import PathExists, IsStickyDir, IsSymLink
 from trashcli.fstab.volume_listing import VolumesListing
 from trashcli.lib.dir_checker import DirChecker
 from trashcli.lib.user_info import UserInfoProvider
